requirements.txt
# requirements.txt
numpy
pandas
tqdm
sentence-transformers
faiss-cpu
transformers
datasets
scikit-learn
xgboost
uvicorn
fastapi
python-multipart
PyPDF2
pdfminer.six
